example-servers/python/fastapi/src/services/custom.py
from fastapi import APIRouter, Request, File, UploadFile, Form
from typing import List, Optional
from pydantic import BaseModel

# placeholder for actual logic
# In a real application, you would integrate with your custom backend service

router = APIRouter()

class ChatMessage(BaseModel):
    role: str
    text: Optional[str] = None
    # Add other fields like 'files' if your chat messages can contain files

class ChatBody(BaseModel):
    messages: List[ChatMessage]

@router.post("/chat")
async def chat(body: ChatBody):
    # Basic echo response for demonstration
    # Replace with your actual chat logic
    last_message = body.messages[-1].text if body.messages else ""
    return {"text": f"Custom service received: {last_message}"}

@router.post("/chat-stream")
async def chat_stream(body: ChatBody):
    # Basic echo response for demonstration
    # Replace with your actual streaming chat logic
    last_message = body.messages[-1].text if body.messages else ""
    # Streaming is more complex and would typically involve something like Server-Sent Events
    # For now, returning a simple response
    return {"text": f"Custom service (stream) received: {last_message}"}

@router.post("/files")
async def files(request: Request, files: List[UploadFile] = File(...)):
    # Basic file handling demonstration
    # Replace with your actual file processing logic
    file_names = [file.filename for file in files]
    # IMPORTANT: You need to decide how to handle file data.
    # Are you saving them? Processing them in memory?
    # For Deep Chat, the response format for files might be different.
    # Refer to https://deepchat.dev/docs/connect/#Response
    # This example just returns the names of the files received.
    return {"message": "Files received successfully by custom service", "files": file_names}
